Re-add config.js for js tests
