+++ test/integration/MultipleClients.test.js
@@ -7,7 +7,7 @@ import { counterId, Defer, pLimitFn } from '../../src/utils'
 import Connection from '../../src/Connection'
 import { StorageNode } from '../../src/stream/StorageNode'
 
-import { clientOptions } from './config'
+import { clientOptions } from './devEnvironment'
 
 const { ControlMessage } = ControlLayer
