+++ test/integration/dataunion/deploy.test.ts
@@ -2,7 +2,7 @@ import { providers } from 'ethers'
 import debug from 'debug'
 
 import { StreamrClient } from '../../../src/StreamrClient'
-import config from '../config'
+import { clientOptions } from '../devEnvironment'
 import { createMockAddress } from '../../utils'
 
 const log = debug('StreamrClient::DataUnion::integration-test-deploy')
@@ -15,7 +15,7 @@ describe('DataUnion deploy', () => {
     let adminClient: StreamrClient
 
     beforeAll(async () => {
-        log(`Connecting to Ethereum networks, config = ${JSON.stringify(config)}`)
+        log('Connecting to Ethereum networks, clientOptions: %o', clientOptions)
         const network = await providerMainnet.getNetwork()
         log('Connected to "mainnet" network: ', JSON.stringify(network))
         const network2 = await providerSidechain.getNetwork()
