+++ test/integration/dataunion/member.test.ts
@@ -2,7 +2,7 @@ import { providers, Wallet } from 'ethers'
 import debug from 'debug'
 
 import { StreamrClient } from '../../../src/StreamrClient'
-import { clientOptions, tokenAdminPrivateKey } from '../config'
+import { clientOptions } from '../devEnvironment'
 import { DataUnion, JoinRequestState } from '../../../src/dataunion/DataUnion'
 import { createMockAddress, expectInvalidAddress, fakePrivateKey } from '../../utils'
 import authFetch from '../../../src/rest/authFetch'
