+++ test/integration/dataunion/transfer.test.ts
@@ -2,7 +2,7 @@ import { BigNumber, Contract, providers, Wallet } from 'ethers'
 import { parseEther } from 'ethers/lib/utils'
 import debug from 'debug'
 import * as Token from '../../../contracts/TestToken.json'
-import { clientOptions, tokenAdminPrivateKey, tokenMediatorAddress, relayTokensAbi } from '../config'
+import { clientOptions, tokenAdminPrivateKey, tokenMediatorAddress, relayTokensAbi } from '../devEnvironment'
 import { getEndpointUrl, until } from '../../../src/utils'
 import { MemberStatus } from '../../../src/dataunion/DataUnion'
 import { StreamrClient } from '../../../src/StreamrClient'
