+++ test/unit/Session.test.ts
@@ -3,7 +3,7 @@ import sinon from 'sinon'
 import { StreamrClient } from '../../src/StreamrClient'
 import { Defer } from '../../src/utils'
 import Session from '../../src/Session'
-import config from '../integration/config'
+import { clientOptions } from '../integration/devEnvironment'
 import { Todo } from '../../src/types'
 
 describe('Session', () => {
